requirements.txt
# ══════════════════════════════════════════════════════════════════════════════
# requirements.txt — AI Interview Platform
# Pinned versions tested and stable on Python 3.11.x
# Install with: pip install -r requirements.txt
# ══════════════════════════════════════════════════════════════════════════════

# ── Web Framework ──────────────────────────────────────────────────────────────
fastapi==0.115.0          # Main web framework for our backend API
uvicorn==0.30.6           # ASGI server that runs our FastAPI app

# ── Templating ─────────────────────────────────────────────────────────────────
jinja2==3.1.4             # Renders HTML templates (our frontend pages)

# ── Groq AI Integration ────────────────────────────────────────────────────────
groq==0.13.0              # Official Groq SDK — 0.13.0 is stable on Python 3.11
httpx==0.27.2             # HTTP client — MUST be pinned to 0.27.x, newer breaks groq

# ── Environment Variable Management ───────────────────────────────────────────
python-dotenv==1.0.1      # Loads .env file so we can store API keys safely

# ── Data Validation ────────────────────────────────────────────────────────────
pydantic==2.9.2           # FastAPI uses this to validate incoming request data

# ── Utilities ─────────────────────────────────────────────────────────────────
anyio==4.4.0              # Async support — pinned to avoid version conflicts
starlette==0.38.6         # FastAPI 0.115.0 requires starlette>=0.37.2,<0.39.0